feat: add globalcontext CEL lib
