+++ pkg/cel/libs/context/mock.go
@@ -34,29 +34,29 @@ func (mock *MockCtx) GetResource(apiVersion, resource, namespace, name string) (
 	return mock.GetResourceFunc(apiVersion, resource, namespace, name)
 }
 
-type mockGctxStore struct {
-	data map[string]store.Entry
+type MockGctxStore struct {
+	Data map[string]store.Entry
 }
 
-func (m *mockGctxStore) Get(name string) (store.Entry, bool) {
-	entry, ok := m.data[name]
+func (m *MockGctxStore) Get(name string) (store.Entry, bool) {
+	entry, ok := m.Data[name]
 	return entry, ok
 }
 
-func (m *mockGctxStore) Set(name string, data store.Entry) {
-	if m.data == nil {
-		m.data = make(map[string]store.Entry)
+func (m *MockGctxStore) Set(name string, data store.Entry) {
+	if m.Data == nil {
+		m.Data = make(map[string]store.Entry)
 	}
-	m.data[name] = data
+	m.Data[name] = data
 }
 
-type mockEntry struct {
-	data any
-	err  error
+type MockEntry struct {
+	Data any
+	Err  error
 }
 
-func (m *mockEntry) Get(_ string) (any, error) {
-	return m.data, m.err
+func (m *MockEntry) Get(_ string) (any, error) {
+	return m.Data, m.Err
 }
 
-func (m *mockEntry) Stop() {}
+func (m *MockEntry) Stop() {}
