+++ pkg/cel/libs/globalcontext/types.go
@@ -1,0 +1,15 @@
+package globalcontext
+
+import (
+	"github.com/google/cel-go/common/types"
+)
+
+var ContextType = types.NewOpaqueType("globalcontext.Context")
+
+type ContextInterface interface {
+	GetGlobalReference(string, string) (any, error)
+}
+
+type Context struct {
+	ContextInterface
+}
